FEA: Make fragments form section queries and generalise
