+++ src/containers/sections/getDatoSectionComponent.jsx
@@ -1,10 +1,8 @@
 import Team from '../../components/sections/Team'
+import JoinTheMovement from '../../components/sections/JoinTheMovement'
 
 const InvalidSection = ({ data }) => {
-  console.warning(
-    'Invalid section id or missing mapping for section data ',
-    data
-  )
+  console.warn('Invalid section id or missing mapping for section data ', data)
   return null
 }
 
@@ -12,6 +10,7 @@ const getDatoSectionComponent = (datoId) => {
   const sectionType = datoId.split('-')[0].replace('DatoCms', '')
   const sectionTypeMap = {
     PeopleSection: Team,
+    JoinSection: JoinTheMovement,
   }
   return sectionTypeMap[sectionType] || InvalidSection
 }
