+++ src/containers/sections/index.js
@@ -1,0 +1,2 @@
+export { default as getDatoSectionComponent } from './getDatoSectionComponent'
+export { default as DatoSections } from './DatoSections'
